Separated test files and moved functions into seperate js files
